+++ src/test/kotlin/kotools/types/number/NegativeIntTest.kt
@@ -511,35 +511,179 @@ class NegativeIntTest {
 
     @Nested
     inner class CompareTo {
+        // ---------- Int ----------
+
         @Test
-        fun `should return 0 when objects have the same value`() {
+        fun `should return 0 with the same int`() {
             // GIVEN
-            val x = NegativeInt(-1)
-            val y = NegativeInt(-1)
+            val x = NegativeInt(0)
+            val y: Int = x.value
             // WHEN
-            val result: Int = x.compareTo(y)
+            val result: Int = x compareTo y
             // THEN
             result assertEquals 0
         }
 
         @Test
-        fun `should return a negative number when this object is less than the other`() {
+        fun `should return a negative number with a greater int`() {
             // GIVEN
-            val x = NegativeInt(-1)
-            val y = NegativeInt(0)
+            val x = NegativeInt(0)
+            val y = 10
             // WHEN
-            val result: Int = x.compareTo(y)
+            val result: Int = x compareTo y
             // THEN
             assertTrue { result < 0 }
         }
 
         @Test
-        fun `should return a positive number when this object is greater than the other`() {
+        fun `should return a positive number with a less int`() {
+            // GIVEN
+            val x = NegativeInt(0)
+            val y = -10
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result > 0 }
+        }
+
+        // ---------- NonZeroInt ----------
+
+        @Test
+        fun `should return 0 with the same non zero int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y: NonZeroInt = x.toNonZeroInt()
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            result assertEquals 0
+        }
+
+        @Test
+        fun `should return a negative number with a greater non zero int`() {
+            // GIVEN
+            val x = NegativeInt(0)
+            val y = NonZeroInt(1)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result < 0 }
+        }
+
+        @Test
+        fun `should return a positive number with a less non zero int`() {
+            // GIVEN
+            val x = NegativeInt(0)
+            val y = NonZeroInt(-1)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result > 0 }
+        }
+
+        // ---------- PositiveInt ----------
+
+        @Test
+        fun `should return 0 with the same positive int`() {
+            // GIVEN
+            val x = NegativeInt(0)
+            val y: PositiveInt = x.toPositiveInt()
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            result assertEquals 0
+        }
+
+        @Test
+        fun `should return a negative number with a greater positive int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y: PositiveInt = -x
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result < 0 }
+        }
+
+        // ---------- StrictlyPositiveInt ----------
+
+        @Test
+        fun `should return a negative number with a greater strictly positive int`() {
+            // GIVEN
+            val x = NegativeInt(0)
+            val y = StrictlyPositiveInt(1)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result < 0 }
+        }
+
+        // ---------- NegativeInt ----------
+
+        @Test
+        fun `should return 0 with the same negative int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y = NegativeInt(x.value)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            result assertEquals 0
+        }
+
+        @Test
+        fun `should return a negative number with a greater negative int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y = NegativeInt(0)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result < 0 }
+        }
+
+        @Test
+        fun `should return a positive number with a less negative int`() {
             // GIVEN
             val x = NegativeInt(-1)
             val y = NegativeInt(-2)
             // WHEN
-            val result: Int = x.compareTo(y)
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result > 0 }
+        }
+
+        // ---------- StrictlyNegativeInt ----------
+
+        @Test
+        fun `should return 0 with the same strictly negative int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y = StrictlyNegativeInt(x.value)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            result assertEquals 0
+        }
+
+        @Test
+        fun `should return a negative number with a greater strictly negative int`() {
+            // GIVEN
+            val x = NegativeInt(-2)
+            val y = StrictlyNegativeInt(-1)
+            // WHEN
+            val result: Int = x compareTo y
+            // THEN
+            assertTrue { result < 0 }
+        }
+
+        @Test
+        fun `should return a positive number with a less strictly negative int`() {
+            // GIVEN
+            val x = NegativeInt(-1)
+            val y = StrictlyNegativeInt(-2)
+            // WHEN
+            val result: Int = x compareTo y
             // THEN
             assertTrue { result > 0 }
         }