+++ changelog.md
@@ -27,6 +27,8 @@ All notable changes to this project will be documented in this file.
   ([#34](https://github.com/kotools/types/issues/34)).
 - Overload comparisons of `kotools.types.number.StrictlyPositiveInt`
   ([#35](https://github.com/kotools/types/issues/35)).
+- Overload comparisons of `kotools.types.number.NegativeInt`
+  ([#36](https://github.com/kotools/types/issues/36)).
 
 ### Changed